python-service/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
Pillow
openai
pytest
pytest-asyncio
httpx
python-dotenv
